// resources/js/game/components/actions/partials/floating-cards/crafting-section/sections/batch-crafting/enums/craft-set-position.ts
export enum CraftSetPosition {
  LEFT_HAND = 'left_hand',
  RIGHT_HAND = 'right_hand',
  BODY = 'body',
  LEGGINGS = 'leggings',
  SLEEVES = 'sleeves',
  GLOVES = 'gloves',
  FEET = 'feet',
  HELMET = 'helmet',
  RING_ONE = 'ring_0',
  RING_TWO = 'ring_1',
  DAMAGE_SPELL = 'spell-damage',
  HEALING_SPELL = 'spell-healing',
}