// resources/js/game/components/actions/partials/floating-cards/crafting-section/sections/crafting/api/definitions/craftable-item-query-definition.ts
export type CraftableItemCraftingType =
  | 'for-class'
  | 'dagger'
  | 'sword'
  | 'claw'
  | 'wand'
  | 'censer'
  | 'stave'
  | 'hammer'
  | 'bow'
  | 'gun'
  | 'fan'
  | 'mace'
  | 'scratch-awl'
  | 'armour'
  | 'ring'
  | 'spell';

export type CraftableItemSubtype =
  | 'shield'
  | 'body'
  | 'leggings'
  | 'sleeves'
  | 'gloves'
  | 'feet'
  | 'helmet'
  | 'spell-damage'
  | 'spell-healing';